move HAL functions to HAL files
